turned playerr.js into backbone model and adjusted tests
