+++ spec/player.spec.js
@@ -1,16 +1,16 @@
-import Player from 'player';
+import Player from 'app/models/player';
 
 describe('Player', function() {
   var testPlayer = new Player();
   describe('initial attributes', function(){
     it ("should have a name attribute of `null` initially", function(){
-      expect(testPlayer.name).toEqual(null);
+      expect(testPlayer.get('name')).toEqual(null);
     });
     it ('should have a mark initally set to `null`', function() {
-      expect(testPlayer.mark).toEqual(null);
+      expect(testPlayer.get('mark')).toEqual(null);
     });
     it ('should have a turn attribute of false', function() {
-      expect(testPlayer.turn).toEqual(false);
+      expect(testPlayer.get('turn')).toEqual(false);
     });
   });
